// src/types/BuildInfo.ts
export type BackendBuildInfo = {
  application: string;
  version: string;
  environment: string;
  branch: string;
  commit: string;
  commitShort: string;
  buildTime: string;
};

export type FrontendBuildInfo = {
  application: string;
  version: string;
  environment: string;
  branch: string;
  commit: string;
  commitShort: string;
  buildTime: string;
};